Widen u16 from_index for u32 GPU synapse buffer
neat-core narrowed SynapseData::from_index from u32 to u16 (neat-core
Issue #177). rust_scorer's GPU mirror SynapseGpu.from_index stays u32
(WGSL has no 16-bit integer type), so the upload in forward_mse_batched
must widen explicitly with u32::from, matching num_synapses. Fixes the
RUSTFLAGS="-D warnings" build break consumed by NEAT-AI-Examples PR #604.

Vibe-Coder-Run-Id: vibe-mqqdqjwm-560862

+++ rust_scorer/src/gpu/forward_mse_batched.rs
@@ -225,7 +225,9 @@ pub fn build_batched_network_data(
         for s in &net.synapses {
             synapses.push(SynapseGpu {
                 weight: s.weight,
-                from_index: s.from_index,
+                // neat-core narrowed SynapseData::from_index to u16 (Issue #177);
+                // widen for the u32 GPU buffer (WGSL has no 16-bit integer type).
+                from_index: u32::from(s.from_index),
             });
         }
